Update from internal repository to sourceforge
git-svn-id: http://svn.code.sf.net/p/bioimagexd/code/bioimagexd/trunk@1478 b2d525f8-dbc2-4032-a3a1-b86af67c733b

+++ vtkBXD/Processing/vtkTestMemory.h
@@ -39,13 +39,13 @@ class VTK_BXD_PROCESSING_EXPORT vtkTestMemory: public vtkObject
   vtkTypeMacro(vtkTestMemory,vtkObject);
   static vtkTestMemory *New();
   virtual void PrintSelf(ostream&, vtkIndent);
-  long CheckMemory();
-  vtkGetMacro(MaxMemory, long);
+  long long CheckMemory();
+  vtkGetMacro(MaxMemory, long long);
 
  protected:
   vtkTestMemory();
   ~vtkTestMemory();
-  long MaxMemory;
+  long long MaxMemory;
 
  private: // Only define operator= and copy constructor to prevent illegal use
   void operator=(const vtkTestMemory&);
